correction event getUserbyID

+++ swagger/spring-server/src/main/java/io/gametown/api/repositories/UserRepository.java
@@ -10,6 +10,6 @@ import java.util.List;
 public interface UserRepository extends CrudRepository<UserEntity, Long>{
     UserEntity findByApplication_ApiKeyAndEmail(String apikey, String email);
   List<UserEntity> findAllByApplication_ApiKeyAndActiveIsTrue(String apiKey);
-  UserEntity findByApplication_ApiKeyAndId(String apiKey, Integer id);
+  UserEntity findByApplication_ApiKeyAndId(String apiKey, Long id);
 
 }
